Mark messages as "trash" when multiple identical messages are detected
Fixes the following problem:

1. An email is sent to foo@foo.example and bar@bar.example,
which are both handled by the same orgmaild.

2. orgmaild forwards to mailhole, where foo.example and bar.example
are separate mailboxes that both have auto-forward enabled.

3. mailhole splits the submitted message by mailbox and attempts to
forward twice.

4. The first attempt to forward goes through, but the second one is
stopped as a detected multiple identical message.

When 4 happens, we shouldn't show the already-forwarded message in the
inbox of the mailbox where it was stopped.

Ideally, orgmaild would split the message by mailbox and submit separate
messages to mailhole, and mailhole's detection of multiple identical
messages would handle this properly.